pyproject.toml
[project]
name = "fastfit-delta"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "psycopg>=3.2.9",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "alembic>=1.16.4",
    "faker>=37.4.2",
    "hypothesis>=6.136.6",
    "psycopg2-binary>=2.9.10",
    "psycopg2[binary]>=2.9.10",
    "pytest>=8.4.1",
    "pytest-alembic>=0.12.1",
    "testcontainers[postgres]>=4.12.0",
]
